Return FedEx's AppliedOptions in meta to the rate caller
FedEx sometimes returns important AppliedOptions in
RateReplyDetails. The specific inspiration for this commit was
that if you want FedEx to ship with One Rate, it will return
the rate for each service both for One Rate AND for non-One Rate.
The service types are the same but the rates are different.
This is a problem because One Rate has some different rules for
packages and you likely need to know if the rate you're looking
at is with One Rate or not. So by looking at the AppliedOptions
returned in meta you can see if 'FEDEX_ONE_RATE' is in there,
and if so, that rate is One Rate.