Update UPS rate test charges and rename surcharge codes
Adjusted the base charge amounts and renamed surcharge codes to 'Fuel surcharge' in UPS rate tests. This change ensures consistency and accuracy in the test data for various UPS services.

changed to use `BaseServiceCharge` instead of `TransportationCharges` because the docs state that `BaseServiceCharge` is The base value of the specific service for the shipment. This is equal to transportation charges - fuel surcharges. and the fuel charge has a separate itemized charge and appears as code 375